requirements.txt
# torch>=1.12.0
# torchvision>=0.13.0
# git+https://github.com/facebookresearch/pytorch3d.git@stable

pytorch-lightning==1.5.5
torchmetrics==0.9.2

opencv-python
matplotlib
chumpy
jupyterlab
scikit-image
av
torch-fidelity
kornia
scipy
omegaconf
ninja
chumpy
V
